Update README.md
Changes are typed into the "Edit file" tab.  Changes can be previewed by clicking the "Preview change" tab.  Changes can be given annotated with a Commit Message in the "Commit changes" portion of the page where they are automatically spell checked. 

Changes can be committed to the current branch or to a new branch.